organize code loosely to the Zend standard with the help of phpcs, minor bug corrections and things improved.
The length of strings are sometimes ignored and it's not perfect. Almost all unnecessary require_once (and all from the forms) changed to require, etc.

Also changes short_open_tags (<?=) to <?php echo which should not be used like this forever (but more on that later) if it's to follow ZF 2 idea of letting use it in the View (and only there).

Signed-off-by: Henrique Vicente de Oliveira Pinto <henvic@henvic-mp.local>